Problem: grids cannot be used with machinic
Solution: add --use-db flag that will connect
to redis and create a glworb: and binary: using
the grid hash and storing the grid thumbnail
along with a an interval that saves the grid
every x seconds. This setup allows grid thumbnails
to appear in a program such as fold-ui and show
updates to the binary.